Refactor: 테스트 코드 리팩토링
