+++ api/src/test/java/com/goodjob/api/controller/member/MemberControllerTest.java
@@ -1,6 +1,5 @@
-package com.goodjob.api.controller;
+package com.goodjob.api.controller.member;
 
-import com.goodjob.api.controller.member.MemberController;
 import com.goodjob.core.domain.member.entity.Member;
 import com.goodjob.core.domain.member.service.MemberService;
 import com.goodjob.core.global.base.redis.RedisUt;
